Quite an update
Basically, a complete "rough" draft of Caitlin and I's paper. (the doctor is not sure yet, on his plan but it is a plan none the less and  almost always will end in saving the world, unless it is his beautiful home that we will never fully see...)
" The sky's burnt orange, with the citadel enclosed in a mighty glass dome, shining under the twin suns. Beyond that, the mountains go on forever. Slopes of deep red grass, capped with snow." - The Doctor